+++ lib/libcithare/password.ml
@@ -38,6 +38,13 @@ module Generate = struct
     all
 
   (**
+    [has_options ~number ~uppercase ~lowercase ~symbole] returns true if at least one of the arguments is true,
+    false otherwise.
+  *)
+  let has_options ~number ~uppercase ~lowercase ~symbole =
+    number || uppercase || lowercase || symbole
+
+  (**
     [create ?(exclude = CharSet.empty) ~number ~uppercase ~lowercase ~symbole] creates a password using several charsets.
     if [number], [uppercase], [lowercase] and [symbole] are all false, [create] defaults to use [alphanumerical] charset
     @raise [Error.CithareError EmptyCharSet] if the char set after all the filter is empty