+++ CHANGELOG.md
@@ -1,3 +1,6 @@
+## [Unrealeased]
+- [cithare-add: Geneate password with the option of cithare-generate-password](https://github.com/EruEri/ocithare/pull/2)
+
 ## [0.9.1]
 - [Clipboard paste other than macOS](https://github.com/EruEri/ocithare/pull/1)